feat(dto): Refactor DTOs to use specific namespaces for Person, Vaccine, and Vaccination
