+++ src/BTG.Vacinacao.Application/DTOs/Vaccination/VaccinationCardDto.cs
