+++ src/BTG.Vacinacao.Application/DTOs/Vaccination/VaccinationRecordDto.cs
